- Move setup sentry inside new module setup_sentry, within a function - Remove not required packages from requirements
